Enh: Drop factory Add: index.js Remove: Node 4 and 5 support
