+++ .travis.yml
@@ -1,12 +1,5 @@
 language: node_js
 node_js:
-  - "4.0"
-  - "4.1"
-  - "4.2"
-  - "4.3"
-  - "4.4"
-  - "4.5"
-  - "4.6"
   - "5.11.0"
   - "6.0"
   - "6.1"
@@ -17,6 +10,8 @@ node_js:
   - "6.6"
   - "6.7"
   - "6.8"
+  - "7.0"
+  - "7.1"
 install:
   - npm install --all
 script:
