+++ index.js
@@ -1,0 +1,2 @@
+module.exports.Cyton = require('./openBCICyton');
+module.exports.Constants = require('./openBCIConstants');
